[AN] refactor: Presentation 유틸리티 함수 공통화 및 패키지 정리 #773
*   [AN] refactor: Color 및 Brush 관련 로직 공통 유틸리티(ColorExtensions, BrushExtensions)로 분리
*   [AN] refactor: CategoryScreen 내 로컬 함수 제거 및 공통 유틸리티 적용
*   [AN] refactor: HorizontalMarginItemDecoration 패키지 경로 이동 및 import 수정
