+++ android/app/src/main/java/com/onair/hearit/presentation/home/adapter/RecommendationCategoryViewHolder.kt
@@ -7,7 +7,7 @@ import com.onair.hearit.databinding.ItemRecommendationCategoryBinding
 import com.onair.hearit.domain.model.RecommendationCategories
 import com.onair.hearit.presentation.HearitClickListener
 import com.onair.hearit.presentation.dpToPx
-import com.onair.hearit.presentation.home.util.HorizontalMarginItemDecoration
+import com.onair.hearit.presentation.home.HorizontalMarginItemDecoration
 
 class RecommendationCategoryViewHolder(
     private val binding: ItemRecommendationCategoryBinding,
