+++ android/app/src/main/java/com/onair/hearit/presentation/home/component/CarouselCard.kt
@@ -29,9 +29,9 @@ import androidx.compose.ui.util.lerp
 import androidx.compose.ui.zIndex
 import com.onair.hearit.R
 import com.onair.hearit.domain.model.RecommendHearit
-import com.onair.hearit.presentation.home.util.toComposeColor
 import com.onair.hearit.presentation.theme.Gray4
 import com.onair.hearit.presentation.theme.HearitTypoGraphy
+import com.onair.hearit.presentation.util.toComposeColor
 import kotlin.math.absoluteValue
 
 @Composable
@@ -58,7 +58,8 @@ fun CarouselCard(
                     scaleX = scale
                     scaleY = scale
                     alpha = alphaValue
-                }.clickable(
+                }
+                .clickable(
                     onClick = onClick,
                     onClickLabel = item.title,
                 ),
